Update patch set 3
Patch Set 3:

> > I have no particular objections to the change but I don't understand the description.  The Go language doesn't have a notion of "right after" except when using a happens-before relationship, but there doesn't seem to be one here.
> 
> This patch only tries to elaborate the side effect of lock release, which can be critical information for some use cases. I think the happens-before relationship as a general language spec still needs more work, especially we should make it explicitly that `sync/atomic` guarantee "sequential consistency", I have commented on https://github.com/golang/go/issues/5045 . 
> 
> Go doesn't have to bring every c++ feature, but memory-model of c++ seems much clearer: https://riptutorial.com/cplusplus/topic/7975/cplusplus11-memory-model?from=timeline&isappinstalled=0

Thanks, but I'm talking about the CL description.  You say "This doc makes it explicit that right after `<-eventCh` in the read side, the `selectnbsend` in the write side is guaranteed to succeed(it won't observe a stale `full()`)."  But as far as I can see there is no happens-before relationship from the write side to the read side in the code that you show.  And the atomics used by the runtime package, in runtime/internal/atomic, are not the same as the atomics used by user programs, in the sync/atomic package.  We shouldn't confuse the two of them.

Patch-set: 3