runtime: use frame pointers to implement runtime.callers()
In order to get this to work, I had to initialize BP to $0 in a bunch of cases (to
signify the end of the stack frames in the case of mcall, clone, etc.), and also
I've already checked in a fix where BP was wrong after nanotime/walltime was
called. There is one remaining problem to fix relating to the BP - the BP needs to
be saved/restored properly in the case of a panic-recover jumping to a deferreturn
(currently only PC/SP are saved/restored).

Unfortunately, there is a bigger problem -- this approach works very well, except
not for cases involving mid-stack inlining. Unfortunately, we need to figure out
all the inline frames at gentraceback/runtime.Callers time, in order to honor the
skip argument and max pcbuf size correctly. But if we need to check for the
inlining func data in runtime.callers(), then the frame-pointer-following
optimization will not be much of a win.

All other tests work except those related to mid-stack inlining and panic-recover
(about 10 tests).

(Some of the initial code grabbed from Keith's first attempt at
https://go-review.googlesource.com/c/go/+/33809 )

DO NOT REVIEW
DO NOT SUBMIT

Updates #16638

Change-Id: I638635c7623e8586935356b0c89f3a842995345a