runtime: use frame pointers to implement gcallers(), speed up trace/mprof
The idea is to grab the stack quickly in gcallers() by following the frame pointer
(when it is in use). For speed, we don't deal with inlining at all in gcallers()
- instead we deal with it in runtime.CallerFrames/Next. The skip argument is hard
to deal with correctly at gcallers time (since we don't do inlining), so we
actually encode the skip argument in the callers array, and do the skip at Next()
time. This does require that the user of gcallers() passes in a large callers
slice.

In order to get this to work, I had to initialize BP to $0 in a bunch of cases (to
signify the end of the stack frames in the case of mcall, clone, etc.), and also
I've already checked in a fix where BP was wrong after nanotime/walltime was
called. There is one remaining problem to fix relating to the BP - the BP needs to
be saved/restored properly in the case of a panic-recover jumping to a deferreturn
(currently only PC/SP are saved/restored).

There's some ugly code to deal with different frame pointer behavior of
systemstack and mcall -- will want to figure out a way to make this cleaner.

All tests pass on linux-amd64

(Some of the initial gcallers code grabbed from Keith's first attempt at
https://go-review.googlesource.com/c/go/+/33809 )

DO NOT REVIEW
DO NOT SUBMIT

Updates #16638

Change-Id: I638635c7623e8586935356b0c89f3a842995345a