cmd/go: adjust heuristics for skipping +incompatible versions
We know of at least one module (github.com/stripe/stripe-go) that has
a run of +incompatible versions, followed by a run of versions with
go.mod files, followed by another run of +incompatible versions.

We want the heuristics for showing +incompatible versions to reflect
the authors' current intent, and it seems clear that the current
intent of the authors of that module is for users to still be on
+incompatible versions. We can detect that intent by checking a single
extra version: the latest release of the latest major version.

In that case, rather than burning time on checking all of the
intervening versions, we will assume that they are all +incompatible.
That means that we may list versions that aren't actually valid, but
since most users won't try to fetch those versions it doesn't really
matter — the improvement in response time outweighs the potential
improvement in precision.

Updates #34165

Change-Id: I5daff3cd2123f94c7c49519babf4eecd509f169e