Update patch set 3
Patch Set 3:

Sorry for the late reply, crazy Q4 end :)

> Patch Set 3:
> 
> What happens if the process is a zombie?


If the process is a zombie, we'll still return it, with its p_stat flag
as C.SZOMB. Our contract for FindProcess says
  "FindProcess looks for a running process by its pid."
but I think if a process is listed by the system e.g. in `ps aux`
we should also be able to return it, or if we decide to, we can filter it out.

Would you suggest that perhaps zombie processes be discarded?

Patch-set: 3
Reviewer: Gerrit User 5137 <5137@62eb7196-b449-3ce5-99f1-c037f21e1705>