Update patch set 1
Patch Set 1:

The use of 32-bit members seems to be intentional and matches struct stat as defined by the kernel (rather than libc). Please see the comment in linux/types.go in x/sys/unix: 

https://go.googlesource.com/sys/+/refs/heads/master/unix/linux/types.go#124

Maybe we should add this type override and comment to package syscall as well.

Patch-set: 1