Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> backport?

I don't think this is worth backporting.
There's an easy workaround. The correct behavior is a panic in any case.
And it seems pretty unlikely to happen in practice. It is a bug in both 1.13 and 1.12, and this is the first report.

Patch-set: 1
Reviewer: Gerrit User 5200 <5200@62eb7196-b449-3ce5-99f1-c037f21e1705>