Update patch set 1
Patch Set 1:

The sync package imports the runtime package, so the runtime package can't import the sync package.  This change can't work as is.  Did you test it yourself?

Also as far as I can tell the basic idea is wrong.  The events variable in netpoll_epoll.go is allocated on the stack.  It is not allocated via malloc.  Using a sync.Pool will cause more memory allocation, not less.  Do you see evidence otherwise?

Patch-set: 1
CC: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>