path/filepath: fix readDirNames to not unconditionally read a directory
`
	names, err := readDirNames(path)
`
readDirNames was always called first to read a directory. And then depending on the value of skipDir, we would stop walking further down. This defeated the purpose of skipDir and can possibly hold on to a large amount of memory if the directory has a large number of files.

We fix this by first checking for the flag and then calling readDirNames.

Change-Id: Idb5633de01ec0077eba37d1e525214b218ee70d6
GitHub-Last-Rev: b3f7c8f02d54d3846eef541b60fd9312336bd4ca
GitHub-Pull-Request: golang/go#36198