cmd/go: convert some testcases to the script framework
This change converts some unparallelized tests to the script
framework so they can run in parallel. The tests are converted
to the script framework instead of being modified because
the script framework tests are easier to read and are self-contained.

Also remove the directory src/cmd/go/testdata/src/badc. The testcase
that used it, "TestDisallowedCSourceFiles" was deleted in
golang.org/cl/19613.

Updates #17751

Change-Id: I0b9b417ae1a9b148067a3e5f8531229f3414f104