cmd/go: convert TestCgoConsistentResults to the script framework
This presents two main challenges. The first is that there's no
way to express tooSlow in the script framework. The second is
that the control flow of a script test can't be determined
by the contents of a file.

I'm not sure what the best approach is for tooSlow. I think the ideal
thing here is to just remove the concept of "tooSlow". Barring
that, if we want to translate the concept to the script framework,
we can either add a [builder] or [canRunTooSlowTests] verb.
(The second one would need a better name of course).

I've picked the option of adding a [builder] verb, but the call to
tooSlow translates to:

    [short] [!builder] skip
    [short] [builder] [!amd64] skip
    [short] [builder] [amd64] [!darwin] [!windows] [!linux] skip

which is awkward. If we take the other route, it would look like

    [short] [!canRunTooSlowTests] skip

which would be simpler.

Updates #17751

Change-Id: I1f3e7664d575219b6fc525bd88babed15d1bd3b2