cmd/go: add [runslow] test to script framework
The tooSlow function is used to create a tier of tests that are
between the full set of tests, and those that are only run in
short mode. This 'middle tier' of tests is run even when
-short is set, but only on the {linux,darwin,windows}/amd64
builders. This guarantees that these tests are always run
on some of the trybots, even though the trybots set -short.

The purpose of this change is to introduce a verb to allow
tests that call tooSlow to be written in the script framework.
At the least, they'd need some way of knowing that they're running
on a builder. This cl adds a [runslow] verb, that when combined
with short has the same effect as a call to tooSlow:

    [short] [!runslow] skip

As alternatives, we could introduce a [builder] condition which
is true if we're running on a builder, or a [slowsystem] verb
which is equivalent to [!runslow].

Change-Id: Ieeaad23a3573ba94f8b1b0d16d8be78971faa4ed