Update patch set 1
Patch Set 1:

(2 comments)

I haven't reviewed the code yet (based on the RET in forwardtail, I assume this is a prototype), but if memmove can shear pointer words, we definitely need to fix it. We've had that bug on other platforms in the past (I'm surprised it would have lasted this long on ARM64).

Patch-set: 1
Reviewer: Gerrit User 5167 <5167@62eb7196-b449-3ce5-99f1-c037f21e1705>