Update patch set 12
Patch Set 12:

Please add the benchmark source to this change.

I'm concerned that calls to netpollBreak are relatively rare, they are already gated by atomic loads, and it's easy to imagine race conditions that cause the poller to not break when it should.  We only call netpollBreak when we find a timer earlier than the one that the poller is waiting for.  In the normal case that can only happen when a new timer is added that triggers earlier than any existing timer, when some P is already sleeping in the poller.  That doesn't seem like the common case to me.  A race condition that causes us to miss a break will cause a timer to not run when it should, and that is always to be avoided.

Patch-set: 12
Reviewer: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>