Update patch set 2
Patch Set 2:

(1 comment)

I've updated the wikipedia link in patch set 2 to be a permanent link based on a suggestion from Michael. It's easy to go from that to the latest page, but the reverse operation is not as easy.

> 1 of 20 TryBots failed:
> Failed on (x/tools) linux-amd64: https://storage.googleapis.com/go-build-log/ec5cdad3/linux-amd64_8d7f7106.log

The TryBot failure is not caused by this change, but it does need to be fixed. I sent CL 213037 that fixes it.

Patch-set: 2
CC: Gerrit User 10033 <10033@62eb7196-b449-3ce5-99f1-c037f21e1705>
Reviewer: Gerrit User 6005 <6005@62eb7196-b449-3ce5-99f1-c037f21e1705>