Update patch set 1
Patch Set 1:

I'm puzzled as to when we use memmove rather than typedmemmove for blocks of memory that contain pointers.

Since there apparently are such cases, would it make sense to provide memmovePtrs and memmoveNoPtrs?  Or is the assumption that it's always going to be better to move memory in pointer-sized chunks anyhow?

Patch-set: 1
CC: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>