Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I'm puzzled as to when we use memmove rather than typedmemmove for blocks of memory that contain pointers.

memmove should never be called directly for memory that contains pointers, but typedmemmove itself calls memmove.

> Since there apparently are such cases, would it make sense to provide memmovePtrs and memmoveNoPtrs?  Or is the assumption that it's always going to be better to move memory in pointer-sized chunks anyhow?

We certainly could do that, but I'm not sure we're really missing out on serious optimization opportunities by making this a general requirements of memmove. Perhaps the benefit of splitting it would be the name would make it more obvious that there's something special.

Patch-set: 1