Update patch set 1
Patch Set 1:

The trybot failure appears to be timeouts. If it was near the edge of its resources, it mightn't be too surprising if this CL tipped it over the edge. The next CL (213703) shrinks everything again, and the same builder is green there.

Patch-set: 1
Reviewer: Gerrit User 5143 <5143@62eb7196-b449-3ce5-99f1-c037f21e1705>