Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> I'm not sure gollvm support buildmode=shared, so this probably doesn't actually matter. Also, I think gollvm driver in general tries to supports flags the same way as gccgo driver (we basically add flags as needed). If this is really needed, we probably want to add it to gollvm instead.

I tested a few cases to check if gollvm support buildmode=shared, it works, but I'm not sure if this build mode is fully supported. Whether gollvm supports shared libraries or not, this is definitely a minor issue if we'll going to support shared library in the future. Originally I made this option a flag of gollvm, but this option doesn't seem to work on shared libraries, see https://linux.die.net/man/1/ld. So we should probably delete this option, I'm not sure, so I made this conservative fix.

Patch-set: 1