Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> There's already a long-path test for RemoveAll(). You can just drop its switch:
> 
>     func TestRemoveAllLongPath(t *testing.T) {
> 	switch runtime.GOOS {
> 	case "aix", "darwin", "dragonfly", "freebsd", "linux", "netbsd", "openbsd", "illumos", "solaris":

This test is not directly usable because ChDir does not support long paths. I spent some time on it but this is not the sole problem. The reported issue and the failing code is using MkDirAll and is fixed with this CL. I suggest to move anything else to a separate issue.

Patch-set: 3