Update patch set 4
Patch Set 4:

I think you misunderstood the bug. To repeat myself, RemoveAll() must check the path argument, and if it's relative construct an absolute path using Getwd(), and pass that to Remove(). There may be code already to evaluate a path for absoluteness, if not consult WinAPI docs; there are many prefixes for absolute paths, not just X:

The test must Chdir(startPath) before calling RemoveAll("foo"). RemoveAll does not return error if the path does not exist, which is why it seemed to work.

https://github.com/golang/go/issues/36375#issuecomment-570759400

Patch-set: 4