Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> I think you misunderstood the bug. To repeat myself, RemoveAll() must check the path argument, and if it's relative construct an absolute path using Getwd(), and pass that to Remove(). There may be code already to evaluate a path for absoluteness, if not consult WinAPI docs; there are many prefixes for absolute paths, not just X:
> 
> The test must Chdir(startPath) before calling RemoveAll("foo"). RemoveAll does not return error if the path does not exist, which is why it seemed to work.
> 
> https://github.com/golang/go/issues/36375#issuecomment-570759400

On my Windows 10 laptop, I can reproduce the hang described in the issue and the fix solves it including removing the relative "foo" directory. The test is valid but not because of RemoveAll(foo) but because of the defer.

The submitted test uses MkDirAll and not MkDir/ChDir/MkDir/... which hits the MAX_PATH limit eventually.

I agree that ChDir should also use fixlongpath but is not required to fix the issue. The related documentation seems to infer that MAX_PATH is active as the limit is around 250 chars. Fixing ChDir is not enough to use a test looping over a 4096 path as I could check.

I really don't mind improving ChDir but it is not related to this issue. Another issue could be created related to the test above including a MAX_PATH check.

Patch-set: 4